+++ src/docker.rs
@@ -7,6 +7,7 @@ use futures::{future, FutureExt};
 use log::error;
 use tokio::runtime::Runtime;
 
+/// Stop container with the label "lazymc.group=group"
 pub fn stop(group: String) {
     debug!(target: "lazymc-docker-proxy::docker", "Stopping containers...");
     let docker: Docker = Docker::connect_with_local_defaults().unwrap_or_else(|err| {
@@ -51,6 +52,7 @@ pub fn stop(group: String) {
     );
 }
 
+/// Start container with the label "lazymc.group=group"
 pub fn start(group: String) {
     debug!(target: "lazymc-docker-proxy::docker", "Starting containers...");
     let docker: Docker = Docker::connect_with_local_defaults().expect("Error connecting to docker");
@@ -92,6 +94,7 @@ pub fn start(group: String) {
     );
 }
 
+/// Stop all containers with the label "lazymc.enabled=true"
 pub fn stop_all_containers() {
     let docker: Docker = Docker::connect_with_local_defaults().unwrap_or_else(|err| {
         error!(target: "lazymc-docker-proxy::docker", "Error connecting to docker: {}", err);
@@ -133,6 +136,7 @@ pub fn stop_all_containers() {
 
 }
 
+/// Get all labels for containers with the label "lazymc.enabled=true"
 pub fn get_container_labels() -> Vec<HashMap<std::string::String, std::string::String>> {
     let docker: Docker = Docker::connect_with_local_defaults().unwrap_or_else(|err| {
         error!(target: "lazymc-docker-proxy::docker", "Error connecting to docker: {}", err);